# app/api/generated/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-07-12T20:02:31+00:00

from __future__ import annotations

from datetime import datetime
from typing import Optional

from pydantic import BaseModel


class RenderDatabase(BaseModel):
    id: Optional[str] = None
    name: Optional[str] = None
    type: Optional[str] = None
    status: Optional[str] = None
    connectionString: Optional[str] = None
    createdAt: Optional[datetime] = None
    updatedAt: Optional[datetime] = None


class RenderDatabaseUser(BaseModel):
    username: Optional[str] = None
    role: Optional[str] = None


class RenderDatabaseConnection(BaseModel):
    host: Optional[str] = None
    port: Optional[int] = None
    database: Optional[str] = None
    user: Optional[str] = None
    password: Optional[str] = None
    uri: Optional[str] = None



class RenderCreateDBRequest(BaseModel):
    name: str
    region: Optional[str] = 'oregon'
    plan: Optional[str] = None
    ownerId: Optional[str] = None
    version: Optional[str] = None


class RenderCreateDBResponse(BaseModel):
    id: Optional[str] = None
    name: Optional[str] = None
    dashboardUrl: Optional[str] = None
    status: Optional[str] = None


class RenderService(BaseModel):
    id: Optional[str] = None
    name: Optional[str] = None
    type: Optional[str] = None
    status: Optional[str] = None
    createdAt: Optional[datetime] = None
    updatedAt: Optional[datetime] = None


class RenderCreateServiceRequest(BaseModel):
    name: str
    type: str


class RenderDeploy(BaseModel):
    id: Optional[str] = None
    serviceId: Optional[str] = None
    status: Optional[str] = None
    createdAt: Optional[datetime] = None
    updatedAt: Optional[datetime] = None


class RenderEnvVar(BaseModel):
    key: Optional[str] = None
    value: Optional[str] = None


class RenderCustomDomain(BaseModel):
    name: Optional[str] = None
    status: Optional[str] = None